+++ test/extensions/persistent_session/plug/cookie_test.exs
@@ -349,11 +349,55 @@ defmodule PowPersistentSession.Plug.CookieTest do
     assert get_from_cache(conn, id, backend: ets) == :not_found
   end
 
-  defp conn_with_session_plug(config) do
-    :get
-    |> Test.conn("/")
+  describe "with telemetry logging" do
+    setup do
+      pid    = self()
+      events = [
+        [:pow, Cookie, :create],
+        [:pow, Cookie, :delete]
+      ]
+
+      :telemetry.attach_many("event-handler-#{inspect pid}", events, fn event, measurements, metadata, send_to: pid ->
+        send(pid, {:event, event, measurements, metadata})
+      end, send_to: pid)
+    end
+
+    test "create and delete", %{conn: new_conn, ets: ets, config: config} do
+      user = %User{id: 1}
+      conn =
+        new_conn
+        |> init_plug(config)
+        |> run_create(user, config)
+
+      assert_receive {:event, [:pow, Cookie, :create], _measurements, %{conn: _conn, key: token, value: {clauses, metadata}}}
+      assert {^clauses, ^metadata} = PersistentSessionCache.get([backend: ets], token)
+      assert clauses == [id: user.id]
+
+      conn
+      |> recycle_session_conn(config)
+      |> init_plug(config)
+      |> run_delete(config)
+
+      assert_receive {:event, [:pow, Cookie, :delete], _measurements, %{conn: _conn, key: deleted_token}}
+      assert deleted_token == token
+      assert PersistentSessionCache.get([backend: ets], deleted_token) == :not_found
+    end
+  end
+
+  defp conn_with_session_plug(config, conn \\ nil) do
+    conn
+    |> Kernel.||(Test.conn(:get, "/"))
     |> PlugSession.call(PlugSession.init(store: ProcessStore, key: "foobar"))
     |> Session.call(Session.init(config))
+  end
+
+  defp recycle_session_conn(old_conn, config) do
+    conn =
+      :get
+      |> Test.conn("/")
+      |> Test.recycle_cookies(old_conn)
+
+    conn_with_session_plug(config, conn)
   end
 
   defp persistent_cookie(conn, cookie_key, id) do