+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 * [`Mix.Tasks.Pow.Extension.Phoenix.Gen.Templates`] `mix pow.extension.phoenix.gen.templates` now dynamically loads template list from the extension base module
 * [`PowResetPassword.Plug`] `PowResetPassword.Plug.load_user_by_token/2` now sets a `:pow_reset_password_decoded_token` key in `conn.private` that will be used in `PowResetPassword.Plug.update_user_password/2`
+* [`Pow.Plug.Session] `:telemetry` events are now dispatched for the following events:
+  * When session is created or renewed in `Pow.Plug.Session.create/3`
+  * When session is deleted in `Pow.Plug.Session.delete/2`
+* [`Pow.Store.Backend.EtsCache`] `:telemetry` events are now dispatched when key value store is deleted, created or invalidated
+* [`Pow.Store.Backend.MnesiaCache`] `:telemetry` events are now dispatched when key value store is deleted, created or invalidated i
 
 ## v1.0.19 (2020-03-13)